Articulating the struggle in doing CQRS
Doing CQRS and Event Sourcing for the first time all on your own is
tough. Sure, you might be a master of your programming language, but if
you're "raised" in a bunch of CRUD models, it can be hard to make the
transition.

Developers can often gravitate towards "frameworks" to help them with
their development. For CRUD, there are quite a few choices (thing Entity
Framework or nHibernate).

In this commit, I am trying to articulate ideas that can "disentangle"
CQRS and Event Sourcing from the development of the Domain, all with a
mind towards better leveraging BDD and TDD style tests that cleanly
distinguishes between testing the Domain (the heart of the app) and
testing how the Domain interacts with the infrastructure that supports
it.